+++ nimbleHMC/R/HMC_samplers.R
@@ -38,9 +38,12 @@
 ## #'     }
 ## #' })
 ## #' 
-## #' N <- 10
-## #' constants <- list(N = N, x = 1:N)
-## #' data <- list(y = 1:N)
+## #' set.seed(0)
+## #' N <- 100
+## #' x <- rnorm(N)
+## #' y <- 1 + 0.3*x + rnorm(N)
+## #' constants <- list(N = N, x = x)
+## #' data <- list(y = y)
 ## #' inits <- list(b0 = 1, b1 = 0.1, sigma = 1)
 ## #' 
 ## #' Rmodel <- nimbleModel(code, constants, data, inits, buildDerivs = TRUE)
@@ -73,6 +76,7 @@ sampler_langevin <- nimbleFunction(
         empirSamp <- matrix(0, nrow = adaptInterval, ncol = d)
         timesRan <- 0
         timesAdapted <- 0
+        ADreset <- 1
         ## checks
         if(!isTRUE(nimbleOptions('enableDerivs')))   stop('must enable NIMBLE derivatives, set nimbleOptions(enableDerivs = TRUE)', call. = FALSE)
         if(!isTRUE(model$modelDef[['buildDerivs']])) stop('must set buildDerivs = TRUE when building model',  call. = FALSE)
@@ -94,7 +98,8 @@ sampler_langevin <- nimbleFunction(
     methods = list(
         jacobian = function(q = double(2)) {
             values(model, target) <<- q[1:d, 1]
-            derivsOutput <- nimDerivs(model$calculate(calcNodes), order = 1, wrt = target)
+            derivsOutput <- nimDerivs(model$calculate(calcNodes), order = 1, wrt = target, reset = ADreset)
+            if(ADreset == 1)   ADreset <<- 0
             grad[1:d, 1] <<- derivsOutput$jacobian[1, 1:d]
             returnType(double(2))
             return(grad)
@@ -117,6 +122,7 @@ sampler_langevin <- nimbleFunction(
             timesAdapted <<- 0
             scaleVec     <<- matrix(1, nrow = d, ncol = 1)
             epsilonVec   <<- scale * scaleVec
+            ADreset <<- 1
         }
     )
 )
@@ -182,22 +188,7 @@ hmc_checkTarget <- function(model, targetNodes, hmcType) {
     dists <- targetDists_unique
     ADok <- rep(TRUE, length(dists))
     for(i in seq_along(dists)) {
-        ##
-        ## if/when modelDef$checkADsupportForDistribution() is added to core nimble,
-        ## change the entire body of this for-loop to instead be:
-        ## ADoak[i] <- model$getModelDef()$checkADsupportForDistribution(dists[i])
-        ##
-        ## these distributions get re-named to a nimble-version, and won't be found:
-        if(dists[i] %in% c('dweib', 'dmnorm', 'dmvt', 'dwish', 'dinvwish'))   next
-        ##
-        ##
-        ## find the function or this distribution:
-        nfObj <- get(dists[i], envir = parent.frame(4))    ## this took a bit of an investigation to make work
-        ## is a user-defined distribution:
-        if(!is.null(environment(nfObj)$nfMethodRCobject)) {
-            ## check for AD support:
-            ADok[i] <- !isFALSE(environment(nfObj)$nfMethodRCobject[['buildDerivs']])
-        }
+        ADok[i] <- model$getModelDef()$checkADsupportForDistribution(dists[i])
     }
     if(!all(ADok))
         stop(paste0(hmcType, ' sampler cannot operate on user-defined distributions which do not support AD calculations.  Try using buildDerivs = TRUE in the definition the distributions: ', paste0(dists[!ADok], collapse = ', ')))
@@ -322,9 +313,12 @@ hmc_setWarmup <- nimbleFunction(
 #'     }
 #' })
 #' 
-#' N <- 10
-#' constants <- list(N = N, x = 1:N)
-#' data <- list(y = 1:N)
+#' set.seed(0)
+#' N <- 100
+#' x <- rnorm(N)
+#' y <- 1 + 0.3*x + rnorm(N)
+#' constants <- list(N = N, x = x)
+#' data <- list(y = y)
 #' inits <- list(b0 = 1, b1 = 0.1, sigma = 1)
 #' 
 #' Rmodel <- nimbleModel(code, constants, data, inits, buildDerivs = TRUE)
@@ -403,6 +397,7 @@ sampler_NUTS_classic <- nimbleFunction(
         sqrtM <- sqrt(M)
         numDivergences <- 0
         numTimesMaxTreeDepth <- 0
+        ADreset <- 1
         ## nimbleLists
         qpNLDef <- nimbleList(q  = double(1), p  = double(1))
         btNLDef <- nimbleList(q1 = double(1), p1 = double(1), q2 = double(1), p2 = double(1), q3 = double(1), n = double(), s = double(), a = double(), na = double())
@@ -464,9 +459,9 @@ sampler_NUTS_classic <- nimbleFunction(
             j <- j + 1
             checkInterrupt()
         }
+        if((timesRan <= nwarmup) & adaptive)   adaptiveProcedure(btNL$a, btNL$na)
         inverseTransformStoreCalculate(qNew)
         nimCopy(from = model, to = mvSaved, row = 1, nodes = calcNodes, logProb = TRUE)
-        if((timesRan <= nwarmup) & adaptive)   adaptiveProcedure(btNL$a, btNL$na)
     },
     methods = list(
         drawMomentumValues = function() {
@@ -495,12 +490,14 @@ sampler_NUTS_classic <- nimbleFunction(
             return(ans)
         },
         gradient_aux = function(qArg = double(1)) {
-            derivsOutput <- nimDerivs(calcLogProb(qArg), order = 1, wrt = nimDerivs_wrt, model = model, updateNodes = nimDerivs_updateNodes, constantNodes = nimDerivs_constantNodes)
+            derivsOutput <- nimDerivs(calcLogProb(qArg), order = 1, wrt = nimDerivs_wrt, model = model, updateNodes = nimDerivs_updateNodes, constantNodes = nimDerivs_constantNodes, reset = ADreset)
+            if(ADreset == 1)   ADreset <<- 0
             returnType(double(1))
             return(derivsOutput$jacobian[1, 1:d])
         },
         gradient = function(qArg = double(1)) {
-            derivsOutput <- nimDerivs(gradient_aux(qArg), order = 0, wrt = nimDerivs_wrt, model = model, updateNodes = nimDerivs_updateNodes, constantNodes = nimDerivs_constantNodes)
+            derivsOutput <- nimDerivs(gradient_aux(qArg), order = 0, wrt = nimDerivs_wrt, model = model, updateNodes = nimDerivs_updateNodes, constantNodes = nimDerivs_constantNodes, reset = ADreset)
+            if(ADreset == 1)   ADreset <<- 0
             grad <<- derivsOutput$value
         },
         leapfrog = function(qArg = double(1), pArg = double(1), eps = double(), first = double(), v = double()) {
@@ -589,6 +586,7 @@ sampler_NUTS_classic <- nimbleFunction(
                             ##for(i in 1:d)   M[i] <<- 1 / warmupCovRegularized[i,i]
                             sqrtM <<- sqrt(M)
                             if(adaptEpsilon) {
+                                inverseTransformStoreCalculate(qNew)   ## defensively ensure model states are up to date
                                 initEpsilon()
                                 epsilonAdaptCount <<- 0
                                 mu <<- log(10 * epsilon)
@@ -710,6 +708,7 @@ sampler_NUTS_classic <- nimbleFunction(
             sqrtM          <<- sqrt(M)
             warmupIntervalNumber <<- 1
             warmupIntervalCount  <<- 0
+            ADreset <<- 1
         }
     ),
     buildDerivs = list(
@@ -787,9 +786,12 @@ treebranchNL_NUTS <- nimbleList(p_beg = double(1), p_end = double(1), rho = doub
 #'     }
 #' })
 #' 
-#' N <- 10
-#' constants <- list(N = N, x = 1:N)
-#' data <- list(y = 1:N)
+#' set.seed(0)
+#' N <- 100
+#' x <- rnorm(N)
+#' y <- 1 + 0.3*x + rnorm(N)
+#' constants <- list(N = N, x = x)
+#' data <- list(y = y)
 #' inits <- list(b0 = 1, b1 = 0.1, sigma = 1)
 #' 
 #' Rmodel <- nimbleModel(code, constants, data, inits, buildDerivs = TRUE)
@@ -877,6 +879,8 @@ sampler_NUTS <- nimbleFunction(
         warningCodes <- array(0, c(max(numWarnings,1), 2))
         numDivergences       <- 0
         numTimesMaxTreeDepth <- 0
+        init_epsilon_next_iter <- FALSE
+        ADreset <- 1
         ## nimbleLists
         treebranchNL <- treebranchNL_NUTS   ## reference input to buildtree
         stateNL <- stateNL_NUTS             ## system state (p, q, H, lp, gr_lp)
@@ -896,7 +900,6 @@ sampler_NUTS <- nimbleFunction(
         if(d  > 1) if(length(M) != d) stop('length of NUTS sampler M must match length of NUTS target nodes', call. = FALSE)
         if(maxTreeDepth < 1) stop('NUTS maxTreeDepth must be at least one', call. = FALSE)
         hmc_checkWarmup(warmupMode, warmup, 'NUTS')
-        init_epsilon_next_iter <- FALSE
     },
     run = function() {
         ## No-U-Turn Sampler based on Stan
@@ -912,12 +915,12 @@ sampler_NUTS <- nimbleFunction(
             if(initializeEpsilon & adaptive)   initEpsilon()
         }
         if(init_epsilon_next_iter) {
-          if(initializeEpsilon)   initEpsilon()
-          Hbar <<- 0
-          logEpsilonBar <<- 0
-          stepsizeCounter <<- 0
-          mu <<- log(10*epsilon)
-          init_epsilon_next_iter <<- FALSE
+            if(initializeEpsilon)   initEpsilon()
+            Hbar <<- 0
+            logEpsilonBar <<- 0
+            stepsizeCounter <<- 0
+            mu <<- log(10*epsilon)
+            init_epsilon_next_iter <<- FALSE
         }
         timesRan <<- timesRan + 1
         if(printTimesRan) print('============ times ran = ', timesRan)
@@ -1005,16 +1008,14 @@ sampler_NUTS <- nimbleFunction(
         accept_prob <- sum_metropolis_prob / n_leapfrog
         copy_state(state_current, state_sample)        ## extraneous copy? could remove?
         ##
-        inverseTransformStoreCalculate(state_sample$q)
-        nimCopy(from = model, to = mvSaved, row = 1, nodes = calcNodes, logProb = TRUE)
         if((timesRan <= nwarmup) & adaptive) {
             if(adaptEpsilon)   adapt_stepsize(accept_prob)
             update <- FALSE
             if(adaptM)   update <- adapt_M()
-            if(update & adaptEpsilon) {
-              init_epsilon_next_iter <<- TRUE
-            }
+            if(update & adaptEpsilon)   init_epsilon_next_iter <<- TRUE
         }
+        inverseTransformStoreCalculate(state_sample$q)
+        nimCopy(from = model, to = mvSaved, row = 1, nodes = calcNodes, logProb = TRUE)
     },
     methods = list(
         copy_state = function(to = stateNL(), from = stateNL()) {
@@ -1040,12 +1041,14 @@ sampler_NUTS <- nimbleFunction(
             return(ans)
         },
         gradient_aux = function(qArg = double(1)) {
-            derivsOutput <- nimDerivs(calcLogProb(qArg), order = 1, wrt = nimDerivs_wrt, model = model, updateNodes = nimDerivs_updateNodes, constantNodes = nimDerivs_constantNodes)
+            derivsOutput <- nimDerivs(calcLogProb(qArg), order = 1, wrt = nimDerivs_wrt, model = model, updateNodes = nimDerivs_updateNodes, constantNodes = nimDerivs_constantNodes, reset = ADreset)
+            if(ADreset == 1)   ADreset <<- 0
             returnType(double(1))
             return(derivsOutput$jacobian[1, 1:d])
         },
         gradient = function(qArg = double(1)) {
-            derivsOutput <- nimDerivs(gradient_aux(qArg), order = 0, wrt = nimDerivs_wrt, model = model, updateNodes = nimDerivs_updateNodes, constantNodes = nimDerivs_constantNodes)
+            derivsOutput <- nimDerivs(gradient_aux(qArg), order = 0, wrt = nimDerivs_wrt, model = model, updateNodes = nimDerivs_updateNodes, constantNodes = nimDerivs_constantNodes, reset = ADreset)
+            if(ADreset == 1)   ADreset <<- 0
             returnType(double(1))
             return(derivsOutput$value)
         },
@@ -1304,6 +1307,7 @@ sampler_NUTS <- nimbleFunction(
             warningInd     <<- 0
             M              <<- Morig
             sqrtM          <<- sqrt(M)
+            ADreset        <<- 1
             ## the adapt_* variables are initialized in before_chain()
             adaptWindow_size    <<- 0
             adapt_initBuffer    <<- 0